# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ferrix API (Ajira Copilot)",
    "version": "1.0.0",
    "summary": "Voice-first worker intake, explainable matching, and the human trust pipeline — as a public API.",
    "description": "Ferrix (Ajira Copilot) public API. Auth contract: mutating routes require 'Authorization: Bearer $FERRIX_API_TOKEN' (or 'x-api-key') when that env var is set; unset = open judge demo. EVERY response carries the 'x-ferrix-auth' header (open-demo | enforced). Hand-maintained; kept in lockstep with docs/API.md and the QA suites.",
    "contact": {
      "name": "Ferrix (Ajira Copilot)",
      "url": "https://github.com/"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev server"
    }
  ],
...
  "paths": {
...
    "/api/chat": {
      "post": {
        "operationId": "postChat",
        "tags": [
          "Chat"
        ],
        "summary": "Send an intake message / start a conversation (guarded mutation)",
        "description": "Single endpoint for the whole intake conversation.\n\n- With `persona` and no `candidateId`: mints a candidate, seeds the persona baseline and the AI intro message.\n- With `candidateId` + `message`: appends the user message, asks the LLM for the next ONE intake question, appends the assistant reply, and returns the full transcript.\n- With `candidateId` and no `message`: returns the transcript without advancing (`mode: \"idle\"`).\n\nNotes: `lang` is accepted as a tolerant alias of `language`. If the LLM is unavailable the route still answers 200 with a scripted reply and `mode: \"fallback\"` (the demo can't die).\n\nThere is no `GET /api/chat` — transcripts arrive in this response or via the case payload.\n\nRate limit: 60 requests/min per IP (429 with `Retry-After`).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transcript + next stage. `mode` is `live` (LLM reply), `fallback` (scripted reply) or `idle` (no message sent).",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, message over 2000 chars / empty after trim, or neither a message nor a persona to create.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Candidate not found (stale/garbage `candidateId`).",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Per-IP rate limit exceeded (60/min).",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the window resets.",
                "schema": {
                  "type": "string"
                }
              },
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/cases/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CaseId"
        }
      ],
      "get": {
        "operationId": "getCase",
        "tags": [
          "Cases"
        ],
        "summary": "Case detail (open read)",
        "description": "Same case shape as the list endpoint, including the candidate's ranked matches (each with human-readable `reasons`) and the append-only case events (actor: AI / Coordinator / System).",
        "security": [],
        "responses": {
          "200": {
            "description": "Case detail.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Case not found.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "operationId": "patchCase",
        "tags": [
          "Cases"
        ],
        "summary": "Advance the trust pipeline (guarded mutation)",
        "description": "Human coordinator action. Enforced state machine (`Invalid transition: …` 400 otherwise):\n\n| action | required current status | new status |\n|---|---|---|\n| `accept` | `new` | `accepted` (auto-assigns mentor, opens verification checklist) |\n| `contact` | `accepted` | `contacted` |\n| `place` | `contacted` | `placed` — requires `opportunityId`; writes a verified PLACEMENT track-record entry |\n| `resolve` | `placed` | `resolved` |\n| `note` | any | unchanged — requires `note` |\n\n`place` on an already-`placed` case and `resolve` on an already-`resolved` case answer `409`. All writes (case status, candidate status, events, track-record entry) run in one transaction — no partial state.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Action applied.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchCaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, unknown action, invalid transition for the current status, `note` missing for `note`, or `opportunityId` missing for `place`.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Case not found, or the `opportunityId` given for `place` does not exist.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Case already placed (`place`) or already resolved (`resolve`).",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "FERRIX_API_TOKEN. Required on mutating endpoints only when the deployment sets FERRIX_API_TOKEN (auth mode `enforced`; watch the `x-ferrix-auth` response header). `x-api-key: <token>` is accepted as an equivalent alternative header. In open demo mode (`x-ferrix-auth: open-demo`) no token is needed."
      }
    },
    "headers": {
      "XFerrixAuth": {
        "description": "Auth mode of the deployment for this request: `open-demo` (mutations open) or `enforced` (mutations require the token). Contract: every response carries it — today it is verifiably present on success responses and 401 auth rejections; the shared 400/404/429 helpers are being wired to stamp it as well.",
        "schema": {
          "type": "string",
          "enum": [
            "open-demo",
            "enforced"
          ]
        }
      }
    },
    "parameters": {
      "CaseId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "Case id (cuid)."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Auth is enforced and the caller did not present the coordinator token (`Authorization: Bearer <token>` or `x-api-key`).",
        "headers": {
          "x-ferrix-auth": {
            "$ref": "#/components/headers/XFerrixAuth"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error.",
        "headers": {
          "x-ferrix-auth": {
            "$ref": "#/components/headers/XFerrixAuth"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Short machine-readable error (first Zod issue message, or a stable label)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Present on 401 auth failures."
          }
        }
      },
...
      "ChatRequest": {
        "type": "object",
        "properties": {
          "candidateId": {
            "type": "string",
            "minLength": 1,
            "description": "Existing candidate. Omit + send persona to mint one."
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "Trimmed by the server; empty-after-trim and >2000 chars are rejected."
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "sw"
            ]
          },
          "lang": {
            "type": "string",
            "enum": [
              "en",
              "sw"
            ],
            "description": "Tolerant alias of `language`."
          },
          "persona": {
            "type": [
              "string",
              "null"
            ],
            "description": "Persona key to mint a candidate from: mama_fua | mama_mboga | boda | cashier | fundi."
          }
        },
        "description": "Either `candidateId`, or `message` and/or `persona`."
      },
      "ChatMessage": {
        "type": "object",
        "required": [
          "id",
          "candidateId",
          "role",
          "content",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "candidateId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ]
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "required": [
          "candidateId",
          "messages",
          "stage",
          "mode"
        ],
        "properties": {
          "candidateId": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          },
          "stage": {
            "type": "string",
            "enum": [
              "greet",
              "work",
              "digital",
              "goal",
              "limits",
              "ready"
            ],
            "description": "Intake stage the next reply addresses."
          },
          "mode": {
            "type": "string",
            "enum": [
              "live",
              "fallback",
              "idle"
            ]
          }
        }
      },
...
      "CaseCandidate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "personaLabel": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "sw"
            ]
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "experience": {
            "type": [
              "string",
              "null"
            ]
          },
          "digitalLiteracy": {
            "type": [
              "string",
              "null"
            ]
          },
          "availability": {
            "type": [
              "string",
              "null"
            ]
          },
          "goal": {
            "type": [
              "string",
              "null"
            ]
          },
          "constraints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "aiSummary": {
            "type": [
              "string",
              "null"
            ]
          },
          "confidence": {
            "type": [
              "number",
              "null"
            ]
          },
          "profile": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/StructuredProfile"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "language",
          "skills",
          "constraints"
        ]
      },
      "CaseMatch": {
        "type": "object",
        "required": [
          "id",
          "score",
          "reasons",
          "opportunity"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "opportunity": {
            "$ref": "#/components/schemas/OpportunityRef"
          }
        },
        "description": "Persisted match as read back from the database. Carries no matcherVersion today (see docs/API.md — serializer is owned by the cases engineer)."
      },
      "CaseEvent": {
        "type": "object",
        "required": [
          "id",
          "actor",
          "action",
          "detail",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "actor": {
            "type": "string",
            "enum": [
              "AI",
              "Coordinator",
              "System"
            ]
          },
          "action": {
            "type": "string",
            "description": "structured | assigned | accepted | contacted | placed | resolved | note | mentor_requested"
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Case": {
        "type": "object",
        "required": [
          "id",
          "ref",
          "priority",
          "status",
          "request",
          "assignedTo",
          "createdAt",
          "updatedAt",
          "candidate",
          "matches",
          "events"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "ref": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "normal",
              "low"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "accepted",
              "contacted",
              "placed",
              "resolved"
            ]
          },
          "request": {
            "type": [
              "string",
              "null"
            ]
          },
          "assignedTo": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "candidate": {
            "$ref": "#/components/schemas/CaseCandidate"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaseMatch"
            }
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaseEvent"
            }
          }
        }
      },
      "CaseResponse": {
        "type": "object",
        "required": [
          "case"
        ],
        "properties": {
          "case": {
            "$ref": "#/components/schemas/Case"
          }
        }
      },
...
      "PatchCaseRequest": {
        "type": "object",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "accept",
              "contact",
              "place",
              "resolve",
              "note"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500,
            "description": "Required for action=note; otherwise becomes the contacted/resolved event detail."
          },
          "opportunityId": {
            "type": "string",
            "minLength": 1,
            "description": "Required for action=place."
          }
        }
      },
      "PatchCaseResponse": {
        "type": "object",
        "required": [
          "ok",
          "caseId",
          "status"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "caseId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "accepted",
              "contacted",
              "placed",
              "resolved"
            ]
          }
        }
      },
...
    }
  }
}